Remove unnecessary providers
