+++ app/src/components/Context/Editor.tsx
@@ -1,0 +1,5 @@
+import { EditorEngine } from '@/lib/editor/engine';
+import { createContext, useContext } from 'react';
+
+const EditorEngineContext = createContext(new EditorEngine());
+export const useEditorEngine = () => useContext(EditorEngineContext);
