+++ app/src/components/Context/Route.tsx
@@ -1,0 +1,5 @@
+import { RouteManager } from '@/lib/routes';
+import { createContext, useContext } from 'react';
+
+const RouteContext = createContext(new RouteManager());
+export const useRouteManager = () => useContext(RouteContext);
