+++ app/src/routes/editor/Canvas/PanOverlay.tsx
@@ -1,7 +1,7 @@
+import { useEditorEngine } from '@/components/Context/Editor';
 import { EditorMode } from '@/lib/models';
 import clsx from 'clsx';
 import { observer } from 'mobx-react-lite';
-import { useEditorEngine } from '..';
 
 interface PanOverlayProps {
     setPosition: React.Dispatch<
