+++ app/src/routes/editor/EditPanel/inputs/BorderInput.tsx
@@ -1,8 +1,8 @@
+import { useEditorEngine } from '@/components/Context/Editor';
 import { ElementStyle, ElementStyleType } from '@/lib/editor/styles/models';
 import { motion } from 'framer-motion';
 import { observer } from 'mobx-react-lite';
 import { useEffect, useState } from 'react';
-import { useEditorEngine } from '../..';
 import ColorInput from './primitives/ColorInput';
 import NumberUnitInput from './primitives/NumberUnitInput';
 import SelectInput from './primitives/SelectInput';
