+++ app/src/routes/editor/EditPanel/inputs/TagDetails.tsx
@@ -1,8 +1,8 @@
+import { useEditorEngine } from '@/components/Context/Editor';
 import { TAG_INFO } from '@/lib/editor/styles/tag';
 import { motion } from 'framer-motion';
 import { observer } from 'mobx-react-lite';
 import { useEffect, useState } from 'react';
-import { useEditorEngine } from '../..';
 
 type TagInfo = {
     title: string;
