+++ app/src/routes/editor/EditPanel/inputs/primitives/NumberUnitInput.tsx
@@ -1,10 +1,10 @@
+import { useEditorEngine } from '@/components/Context/Editor';
 import { constructChangeCurried } from '@/lib/editor/styles/inputs';
 import { ElementStyle } from '@/lib/editor/styles/models';
 import { parsedValueToString, stringToParsedValue } from '@/lib/editor/styles/numberUnit';
 import { ChevronDownIcon } from '@radix-ui/react-icons';
 import { observer } from 'mobx-react-lite';
 import React, { ChangeEvent, useEffect, useState } from 'react';
-import { useEditorEngine } from '../../..';
 
 const NumberUnitInput = observer(
     ({
