+++ app/src/routes/editor/EditPanel/inputs/primitives/SelectInput.tsx
@@ -1,3 +1,4 @@
+import { useEditorEngine } from '@/components/Context/Editor';
 import { ToggleGroup, ToggleGroupItem } from '@/components/ui/toggle-group';
 import { constructChangeCurried } from '@/lib/editor/styles/inputs';
 import { ElementStyle } from '@/lib/editor/styles/models';
@@ -14,7 +15,6 @@ import {
 } from '@radix-ui/react-icons';
 import { observer } from 'mobx-react-lite';
 import { useEffect, useState } from 'react';
-import { useEditorEngine } from '../../..';
 
 const OVERRIDE_OPTIONS: Record<string, string> = {
     'flex-start': 'start',
