+++ app/src/routes/editor/EditPanel/inputs/primitives/TextInput.tsx
@@ -1,8 +1,8 @@
+import { useEditorEngine } from '@/components/Context/Editor';
 import { constructChangeCurried } from '@/lib/editor/styles/inputs';
 import { ElementStyle } from '@/lib/editor/styles/models';
 import { parsedValueToString, stringToParsedValue } from '@/lib/editor/styles/numberUnit';
 import { appendCssUnit } from '@/lib/editor/styles/units';
-import { useEditorEngine } from '@/routes/editor';
 import { observer } from 'mobx-react-lite';
 import React, { useEffect, useState } from 'react';
