+++ app/src/routes/editor/TopBar/SharePopover/index.tsx
@@ -1,9 +1,9 @@
+import { useEditorEngine } from '@/components/Context/Editor';
 import { Button } from '@/components/ui/button';
 import { Label } from '@/components/ui/label';
 import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover';
 import { GlobeIcon, Share2Icon } from '@radix-ui/react-icons';
 import { useEffect, useState } from 'react';
-import { useEditorEngine } from '../..';
 import { MainChannels } from '/common/constants';
 import { TunnelResult } from '/common/models/tunnel';
